Этот пул-реквест вносит существенные изменения в способ обработки идентификаторов музыкальных групп и связанных с ними объектов, заменяя Integer на Long во всей кодовой базе и схеме базы данных. Это повышает масштабируемость и согласованность при работе с большими наборами данных. Кроме того, в службу аутентификации добавлен новый метод проверки прав пользователей. Ниже приведены наиболее важные изменения, сгруппированные по темам.

Миграция типа идентификатора (Integer → Long):

Изменены все идентификаторы музыкальных групп и связанных с ними сущностей с Integer на Long в классах моделей, DTO, интерфейсах/реализациях сервисов, контроллерах, репозиториях и тестовых сценариях (MusicBand, BestBandAwardRequest и т. д.), что обеспечивает единообразие во всём приложении. [1] [2] [3] [4] [5] [6] [7] [8] [9] [10] [11] [12] [13] [14] [15] [16] [17] [18] [19] [20] [21]

Обновлена схема базы данных: столбцы music_band.id и best_band_award.band_id изменены с SERIAL/INTEGER на BIGSERIAL/BIGINT для лучшей поддержки больших значений идентификаторов. [1] [2]

Аутентификация и разрешения:

Добавлен новый метод authenticatedUserPermission(Permission permission) в UserService и его реализация, позволяющие проверять разрешения для пользователя, прошедшего аутентификацию в данный момент. [1] [2]
Настройка репозитория и сервисов:

Обновлены интерфейсы репозитория и методы обслуживания, теперь для идентификаторов используется Long вместо Integer, в том числе для операций запроса и удаления. [1] [2] [3]
Обновления контроллера:

Изменены параметры конечной точки контроллера и тела запросов для поддержки Long идентификаторов, что обеспечивает совместимость с обновлёнными уровнями сервиса и репозитория. [1] [2]
Рефакторинг тестов:

Мы провели рефакторинг всех связанных тестов, чтобы использовать идентификаторы Long и сохранить тестовое покрытие, а также обеспечить корректность работы после миграции. [1] [2] [3] [4] [5] [6] [7]

+++ src/main/java/org/is/bandmanager/controller/MusicBandController.java
@@ -43,7 +43,7 @@ public class MusicBandController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<MusicBandDto> getMusicBand(@PathVariable Integer id) {
+    public ResponseEntity<MusicBandDto> getMusicBand(@PathVariable Long id) {
         MusicBandDto musicBand = musicBandService.get(id);
         return ResponseEntity.ok(musicBand);
     }
@@ -80,26 +80,26 @@ public class MusicBandController {
 
     @PutMapping("/{id}")
     public ResponseEntity<MusicBandDto> updateMusicBand(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @Valid @RequestBody MusicBandRequest request) {
         MusicBandDto updatedMusicBand = musicBandService.update(id, request);
         return ResponseEntity.ok(updatedMusicBand);
     }
 
     @PutMapping("/{id}/remove-participant")
-    public ResponseEntity<MusicBandDto> removeParticipant(@PathVariable Integer id) {
+    public ResponseEntity<MusicBandDto> removeParticipant(@PathVariable Long id) {
         MusicBandDto updatedBand = musicBandService.removeParticipant(id);
         return ResponseEntity.ok(updatedBand);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<MusicBandDto> deleteMusicBand(@PathVariable Integer id) {
+    public ResponseEntity<MusicBandDto> deleteMusicBand(@PathVariable Long id) {
         MusicBandDto deletedMusicBand = musicBandService.delete(id);
         return ResponseEntity.ok(deletedMusicBand);
     }
 
     @DeleteMapping()
-    public ResponseEntity<List<MusicBandDto>> deleteMusicBands(@RequestBody List<Integer> ids) {
+    public ResponseEntity<List<MusicBandDto>> deleteMusicBands(@RequestBody List<Long> ids) {
         List<MusicBandDto> deletedMusicBands = musicBandService.delete(ids);
         return ResponseEntity.ok(deletedMusicBands);
     }
